Implement permissions for Activity and Add BigDecimal for price

+++ src/main/java/com/prime/rushhour/domain/activity/dto/ActivityRequest.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-import java.time.Duration;
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ActivityRequest(
@@ -15,7 +15,7 @@ public record ActivityRequest(
         String name,
         @NotNull(message = "Price is a mandatory field")
         @Min(value = 0, message = "Price must be greater than 0")
-        Double price,
+        BigDecimal price,
         @NotNull(message = "Duration is a mandatory field")
         @Min(value = 0, message = "Duration must be greater than 0")
         Long duration,
